+++ config.properties
@@ -1,9 +1,9 @@
 project=Interepid
 
 #list of application schema prefixes 
-#application.list=windows,portal,tablet
-application.list=tablet
+application.list=windows,portal,tablet
+#application.list=tablet
 
-template=../data/template.pptx
+template=../data/Tech DD Findings - Template v1.2.pptx
 output.folder=./deck
